Bug fixes and tweaks

+++ Source/ColourUtility.cpp
@@ -176,9 +176,9 @@ namespace ColourUtility
 
 		if (brightness != 100)
 		{
-			r = std::round((brightness / 100) * r);
-			g = std::round((brightness / 100) * g);
-			b = std::round((brightness / 100) * b);
+			r = std::round(((double)brightness / 100) * (double)r);
+			g = std::round(((double)brightness / 100) * (double)g);
+			b = std::round(((double)brightness / 100) * (double)b);
 		}
 
 		switch (colourspace)
